enterprise_finance/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
Django==5.2
django-axes==7.1.0
django-tinymce==4.1.0
et_xmlfile==2.0.0
idna==3.10
narwhals==1.37.1
oauthlib==3.2.2
openpyxl==3.1.5
packaging==25.0
plotly==6.0.1
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
urllib3==2.4.0
